+++ CHANGELOG.md
@@ -17,6 +17,7 @@ This changelog also contains important changes in dependencies.
 - Allow `feComposite` k1-4 coefficients to be larger than 1.0
   This matches browsers behaviour.
 - Use `flate2` instead of `libflate` for GZip decoding.
+- (usvg) `fill` and `stroke` attributes will always be set for `path` now.
 
 ### Fixed
 - `feComposite` with fully transparent regions was producing invalid results.